telegram_bot/requirements.txt
openai==1.12.0
psycopg2-binary==2.9.9
python-telegram-bot==20.7
reportlab==4.0.7
tabulate==0.9.0
